# docs/openapi/gsf-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GSF API",
    "version": "0.1.1",
    "description": "Public HTTP surface of GSF, served by the Next.js route handlers. Generated by `pnpm openapi` — do not edit by hand."
  },
  "paths": {
...
    "/api/connections": {
      "get": {
        "operationId": "get_api_connections",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse_PublicConnection_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No authenticated caller could be resolved"
          },
          "403": {
            "description": "The caller lacks the required permission"
          }
        },
        "summary": "connectionsApi.getAll — list connections (viewable by all).",
        "tags": ["connections"],
        "description": "connectionsApi.getAll — list connections (viewable by all).",
        "x-gsf-source": "frontend/app/api/connections/route.ts",
        "x-gsf-permissions": {
          "connection": ["read"]
        },
        "x-gsf-proxied": true
      },
      "post": {
        "operationId": "post_api_connections",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_PublicConnection_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No authenticated caller could be resolved"
          },
          "403": {
            "description": "The caller lacks the required permission"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "connectionsApi.create — add a connection (admin only; carries credentials).",
        "tags": ["connections"],
        "description": "connectionsApi.create — add a connection (admin only; carries credentials).",
        "x-gsf-source": "frontend/app/api/connections/route.ts",
        "x-gsf-permissions": {
          "connection": ["manage"]
        },
        "x-gsf-proxied": true
      }
    },
    "/api/connections/{database_name}": {
      "delete": {
        "operationId": "delete_api_connections_database_name_",
        "parameters": [
          {
            "description": "Catalog database name, which doubles as the connection's identity.",
            "in": "path",
            "name": "database_name",
            "required": true,
            "schema": {
              "description": "Catalog database name, which doubles as the connection's identity.",
              "title": "Database Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_PublicConnection_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No authenticated caller could be resolved"
          },
          "403": {
            "description": "The caller lacks the required permission"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "connectionsApi.delete — remove a connection (admin only).",
        "tags": ["connections"],
        "description": "connectionsApi.delete — remove a connection (admin only).",
        "x-gsf-source": "frontend/app/api/connections/[database_name]/route.ts",
        "x-gsf-permissions": {
          "connection": ["manage"]
        },
        "x-gsf-proxied": true
      }
    },
    "/api/connections/source": {
      "get": {
        "description": "connectionsApi.isEnvSource — whether connections come from env config.",
        "operationId": "get_api_connections_source",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Is Env Source Api Connections Source Get",
                  "type": "boolean"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No authenticated caller could be resolved"
          },
          "403": {
            "description": "The caller lacks the required permission"
          }
        },
        "summary": "connectionsApi.isEnvSource — whether connections come from env config.",
        "tags": ["connections"],
        "x-gsf-source": "frontend/app/api/connections/source/route.ts",
        "x-gsf-permissions": {
          "connection": ["read"]
        },
        "x-gsf-proxied": true
      }
    },
    "/api/connections/test": {
      "post": {
        "description": "connectionsApi.test — validate a connection config (admin only).",
        "operationId": "post_api_connections_test",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionTestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No authenticated caller could be resolved"
          },
          "403": {
            "description": "The caller lacks the required permission"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "connectionsApi.test — validate a connection config (admin only).",
        "tags": ["connections"],
        "x-gsf-source": "frontend/app/api/connections/test/route.ts",
        "x-gsf-permissions": {
          "connection": ["manage"]
        },
        "x-gsf-proxied": true
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConnectionBody": {
        "properties": {
          "connection": {
            "additionalProperties": true,
            "title": "Connection",
            "type": "object"
          }
        },
        "required": ["connection"],
        "title": "ConnectionBody",
        "type": "object"
      },
      "ConnectionTestResponse": {
        "additionalProperties": true,
        "description": "``schemas`` is empty for connectors that cannot enumerate schemas.",
        "properties": {
          "schemas": {
            "items": {
              "type": "string"
            },
            "title": "Schemas",
            "type": "array"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": ["success", "schemas"],
        "title": "ConnectionTestResponse",
        "type": "object"
      },
...
      "DataResponse_PublicConnection_": {
        "additionalProperties": true,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PublicConnection"
          }
        },
        "required": ["data"],
        "title": "DataResponse[PublicConnection]",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ListResponse_PublicConnection_": {
        "additionalProperties": true,
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/PublicConnection"
            },
            "title": "Data",
            "type": "array"
          }
        },
        "required": ["data", "count"],
        "title": "ListResponse[PublicConnection]",
        "type": "object"
      },
...
      "PublicConnection": {
        "additionalProperties": true,
        "description": "A UI-managed connection with its credentials stripped.\n\n``connection`` is absent from the delete response, which only echoes the\ndatabase name it tore down.",
        "properties": {
          "connection": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Connection"
          },
          "database_name": {
            "title": "Database Name",
            "type": "string"
          }
        },
        "required": ["database_name"],
        "title": "PublicConnection",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  }
}